requirements.txt
Django==3.2
psycopg2
ipython
docutils
django-environ
djangorestframework
celery
django-celery-results
django-rest-swagger
gunicorn